+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Better scaling of lines (ROI, alginment grid etc.)
 - Accidental scrolling on image rotation slider without focus is not longer possible
 - The marker box in recognition now no longer reacts to mouse wheel scrolling
+- ROI `show` set to false now makes the `fix` ROI button unable to toggle
 - Fix: roi not shown on startup
 - Fix: wrong margin for some UI boxes
 - Fix: code marker ID 0 was not displayed